payment_gateway/pyproject.toml
[project]
name = "payment-gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.13",
    "uvicorn>=0.25.0",
    "python-dotenv>=1.0.0",
    "confluent-kafka[avro, schema-registry]>=2.10.1",
    "authlib>=1.6.0",
    "redis>=6.2.0",
    "python-dotenv>=1.0.0",
    "aioredis>=2.0.1",
    "sqlalchemy>=2.0.36",
    "pydantic>=2.7.2",
    "asyncpg>=0.30.0"
]
